RND machines now use connect_techweb (automatically updates) (#71420)
## About The Pull Request

``connect_techweb`` handles registering signals to the techweb it
connects to for automatic updating on node researching, which is what is
used when you multitool to connect something to it.
However, this wasn't used on its Initialize, meaning you had to connect
it manually with a multitool to get the automatic updating functioning
properly. This fixes that.

## Why It's Good For The Game

Closes https://github.com/tgstation/tgstation/issues/71418
Techfabs work as intended now.

## Changelog

:cl:
fix: Techfabs now automatically update to researched designs again.
/:cl:
# Conflicts:
#	code/modules/research/rdmachines.dm