Improved nightshift check on security level change (#79369)
## About The Pull Request

Modifies the security level change proc to only fire the nightshift
subsystem if the subsystem is running, and if a lighting change is
required. If a lighting change is required, nightshift has its next fire
set 7 seconds after, so announcements and sounds don't stack.

## Why It's Good For The Game

Nightshift is only updated when required, eliminating nuisance
announcements when levels unrelated to red/delta are selected.

## Changelog

:cl: LT3
code: Changing security levels will only trigger the nightshift
subsystem if lighting changes are required
/:cl: