Bumps rust-g to `3.0.0`  (#76663)
## About The Pull Request

You can read more about what's introduced in the last four versions
here:

https://github.com/tgstation/rust-g/releases

The immediate effects for us are as follows:
Closes #76549
Closes #76514
Closes #76502
Closes #76490
Closes #76398
Closes #76798
Closes #76921
Closes #76934
Closes #76950

the updates to install rust_g.sh were to cut down on the copypasta code
and prevent desyncs.

:cl:
code: The currently operating rust-g version on a live server is posted
to places like the runtime.log, in the same place where the revision
information and any applicable test merges already were.
/:cl:
# Conflicts:
#	.github/workflows/run_integration_tests.yml
#	code/__DEFINES/rust_g.dm
#	rust_g.dll
#	tools/tgs_scripts/PreCompile.sh