[PORT] Fix ByondUIs Not Cleaning Up (#552)
## About The Pull Request

https://github.com/BeeStation/BeeStation-Hornet/pull/10512

This can build up to cause insane lag. Let's not.

:cl:
fix: An issue where ByondUIs could stack up and lag.
/:cl:

<!-- Both :cl:s are required for the changelog to work! You can put your
name to the right of the first :cl: if you want to overwrite your GitHub
username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->