[Port] Maintenance vent soulening (#567)
## About The Pull Request

This pr is a port of the changes to vents in this PR:
- https://github.com/NovusSS13/NovusSS13/pull/157

This PR adds a cool distortion effect to the air above maintenance
vents. The effect goes away when the vent is turned off. I'm only adding
this for soul reasons and because the new map #562 has alot of these and
it would add more soul to the map.

## How Does This Help ***Gameplay***?
Cooler things to look at during the gaming.

## How Does This Help ***Roleplay***?
Immersion.

## Proof of Testing

<!-- Include any screenshots/videos/debugging steps of the code
functioning successfully, between the </summary> and </details> code
blocks. -->
<!-- To our mappers and spriters: Posting screenshots of content INSIDE
EDITORS (aseprite, PDN, SDMM, ect) is NOT valid proof of testing. Please
make sure that you COMPILE the game and provide PROOF you tested your
edits. -->
<!-- New content PRs will not be merged without screencaps of what every
added thing looks like in game. -->

<details>
<summary>Screenshots/Videos</summary> <!-- Leave the line after this one
empty. Embeds like breaking if you don't -->


https://github.com/Artea-Station/Artea-Station-Server/assets/79924768/14f326ec-b078-43f3-92fc-bb751e44160a


</details>

## Changelog
:cl:
add: Maintenance vents now have a cool distortion effect for the air
around them!
/:cl: