Overmap resprite (#606)
## About The Pull Request

I totally didn't forget that this branch existed until looking at it
last week.

Anyways, this does what it says on the title. This PR is a resprite of
the overmap to fit in better with the rest of our UI's. Also did this
because the normal overmap is boring as hell and the sprites are kind of
ass. Overmap hazards now have also been given slight animations or have
been replaced with a completely animated sprite, see testing video for
what im talking about.

Credit goes to shiptest for the carp sprites and inspiration, Thanks
guys.
-
https://github.com/shiptest-ss13/Shiptest/blob/master/icons/misc/overmap.dmi

## How Does This Help ***Gameplay***?
Piloting the ship is so much awesomer now dude, trust me, you're not
drowning in a sea of blue anymore either because the background is now
this cool black gradient thing wow!

## How Does This Help ***Roleplay***?
Minimal impact on RP.

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary> <!-- Leave the line after this one
empty. Embeds like breaking if you don't -->

## Before


https://github.com/user-attachments/assets/874b0935-fc59-42b8-9ea8-e04c2a9d3be0

## After


https://github.com/user-attachments/assets/87daa519-3681-4d66-92cc-3133c7ea92cc

</details>

## Changelog
:cl:
imageadd: Resprited the overmap and the associated objects/hazards.
/:cl: