[PORT] The recoilening  (#589)
## About The Pull Request
This pr is a port of: 
- https://github.com/Mojave-Sun/mojave-sun-13/pull/430

This PR significantly improves recoil (screenshake when shooting guns)
because our current recoil sucks ass. Instead of a simple screenshake
its now based on the direction you're shooting in. I have also gone
ahead and added recoil values to the guns we have in-game right now.

Ballistic guns have a recoil value of `1`
Energy guns have a recoil value of `0.2`
Disablers and the Miniature energy gun have a recoil value of because
they're tiny `0.1`

This is yet another expansion of the guns and other cool shit pr's like:
#520 #455 and #460.

## How Does This Help ***Gameplay***?
Feedback to in game actions is always good and guns not having recoil
when being shot is just weird.

## How Does This Help ***Roleplay***?
Immersion probably

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary> <!-- Leave the line after this one
empty. Embeds like breaking if you don't -->

## Old Recoil


https://github.com/Artea-Station/Artea-Station-Server/assets/79924768/3a483a6d-a98f-4ae7-a34e-f4b7750ac7c1

## New Recoil


https://github.com/Artea-Station/Artea-Station-Server/assets/79924768/6bcefbc4-3070-44f6-90ba-45638c99b553

## New recoil on a disabler


https://github.com/Artea-Station/Artea-Station-Server/assets/79924768/7ac9e71a-eebb-4be8-afc3-a971c3d2d013



</details>

## Changelog
:cl:
qol: Recoil is alot less awful now and is based on the direction you're
shooting in.
/:cl: