[PORT] Droning Subsystem (#596)
## About The Pull Request

This PR ports the Droning Subsystem from EFN.

The Droning SS is used for playing looped music for the area's that
players are in. Different music can be given to different areas and
areas can have multiple different tracks tied to them that will change
when a player enters said area. In my opinion this is a super cool
Subsystem that we could benefit a lot from because it adds much more
atmosphere to the ship when properly setup. (Video previews in proof of
testing are a bit short due to github file size limitations)

This Subsystem does not replace the already existing ship ambience,
instead it plays at the same time adding to the soundscape of the server
in a unique way.

Currently on this PR I have given the different Areas featured on
Belryth their own tracks and it is making for a nice and dark atmosphere
aboard an old and forgotten ship. All the tracks on the PR currently are
simple liminal tracks and
ambient droning. This SS does not come with a pref that players can
toggle and I am not intending to add one either.

All tracks have been sourced from: https://freesound.org/ See below for
all the tracks that have been used on this PR.

<details>
<summary>Tracks used:</summary> <!-- Leave the line after this one
empty. Embeds like breaking if you don't -->

https://freesound.org/people/ZHR%C3%98/sounds/692011/

https://freesound.org/people/ZHR%C3%98/sounds/531447/

https://freesound.org/people/ZHR%C3%98/sounds/692006/

https://freesound.org/people/ZHR%C3%98/sounds/692009/

https://freesound.org/people/ZHR%C3%98/sounds/702966/

https://freesound.org/people/Mellau/sounds/505667/

https://freesound.org/people/GregorQuendel/sounds/674408/

</details>

Code and inspiration taken from Escape from Nevado:
- https://gitgud.io/the-itobe-denomination/escape-from-nevado
-
https://gitgud.io/the-itobe-denomination/escape-from-nevado/-/blob/master/modular_septic/code/controllers/subsystem/droning.dm

## How Does This Help ***Gameplay***?
Background music ontop of the normal sounds + the ambience sound system
is so much better. The atmosphere is completely different on the ship
now and it feels much darker and drearier.

## How Does This Help ***Roleplay***?
Immersion, adding this to our soundscape makes me feel 10x more immersed
when im standing in some shitty dark corner on Belryth.

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary> <!-- Leave the line after this one
empty. Embeds like breaking if you don't -->

# SOUND ON 🔉 


https://github.com/Artea-Station/Artea-Station-Server/assets/79924768/dc8e2aa8-9b01-482d-9604-889805cf91fa

## Normal ship droning


https://github.com/Artea-Station/Artea-Station-Server/assets/79924768/5c5da108-8648-4068-a27a-8b014c193940

## Medical Droning


https://github.com/Artea-Station/Artea-Station-Server/assets/79924768/293a0280-5cb7-42a7-946a-b10474037a12

## Engi Droning


https://github.com/Artea-Station/Artea-Station-Server/assets/79924768/ef6a093f-e4e2-47d1-97d2-cbd358023ca9

</details>

## Changelog
:cl:
soundadd: Added multiple tracks of background music that change
depending on the area of the ship you are in.
/:cl:

---------

Signed-off-by: Rimi Nosha <riminosha@gmail.com>
Co-authored-by: Rimi Nosha <riminosha@gmail.com>